Better example: good variable names, some comments, reorganization.
Don't use a global array, indexed at [i*100] for partial summing.
It was completely non-obvious why the array needed to be 2000 items
large for only 8 threads, and accessed at [i*100] (this was to
avoid false sharing). Instead, sum into a stack variable and write
only once to a packed, 8-element array.